Rename `/functions/` to `/modules/`
